Revert "Merge pull request #2206 from hmcts/NFDIV-2207"
This reverts commit b1538958c8e942fd8a5cdb1c44b2e5020d12562c, reversing
changes made to 2d692f001cb97c71c79852daae17519e8c285b23.
